+++ src/main/kotlin/com/mpi/kissing_company/repositories/RoleRepository.kt
@@ -3,6 +3,6 @@ package com.mpi.kissing_company.repositories
 import com.mpi.kissing_company.entities.Role
 import org.springframework.data.jpa.repository.JpaRepository
 
-internal interface RoleRepository : JpaRepository<Role?, Int>{
+internal interface RoleRepository : JpaRepository<Role?, Long>{
     fun findByName(name: String): Role
 }
